+++ src/vm-repair/HISTORY.rst
@@ -5,6 +5,11 @@ Release History
 ++++++
 Add `--encrypt-recovery-key` string parameter to `vm repair create` to use recovery key provided by the user to unlock the disk for a confidential VM. 
 
+1.0.9
+++++++
+Fixed and updated several vm-repair tests for better coverage. 
+Removed and updated broken image aliases pointing at images that no longer existed. 
+
 1.0.8
 ++++++
 SELFHELP telemetry added as initiator. Extra parameters is introduced at the backend to capture the telemetry data.